fix : runrecord 속성 추가, 변경 (domain layer)

+++ RunWithMe/domain/src/main/java/com/side/domain/model/RunningInfo.kt
@@ -1,7 +1,7 @@
 package com.side.domain.model
 
 data class RunningInfo(
-    val challengeSeq: Int,
+    val challengeSeq: Long,
     val challengeName: String,
     val goalAmount: Long,
     val goalType: Int
